First commit :v:
We still have a choice today: nonviolent coexistence or violent
coannihilation. We must move past indecision to action. We must find
new ways to speak for peace in Vietnam and justice throughout the
developing world, a world that borders on our doors. If we do not act,
we shall surely be dragged down the long, dark, and shameful corridors
of time reserved for those who possess power without compassion, might
without morality, and strength without sight.

– Martin Luther King Jr.