Create action items
This update adds a new `ActionItem` model, along with an
administration area to add, update and delete action items.

Action items belong to both Issues and Categories, and consist
of the following attributes:

Attribute Name    Description                         Type    Default
-----------------------------------------------------------------------
`body`            A long-form description of the     `text`    `NULL`
                  action.
`event`           If this action is an event.        `boolean`  false
`eyebrow`         A two-or-three word description,   `string`  `NULL`
                  such as _“Protest for…”_, or
                  _“Call Congress”_.
`happening_at`    When this action takes place.      `datetime` `NULL`
`image_data`      A large image, used for feature    `text`    `NULL`
                  backgrounds and heroes.
`location`        The location of this action.       `string`   `NULL`
`priority`        An enum representing the perceived `integer`  0
                  priority of the action (_Do It_,
                  _Urgent_, _Do or Die!_).
`summary`         A short text summary of what the   `text`    `NULL`
                  action is about.
`time_commitment` The estimated time it will take to `string`  `NULL`
                  perform the action.
`title`           The title of the action item.      `string`  required

Closes #8