Rename action_item to action
Prior to this update, actions had been named “action items”. The reason
for this is that a conflict exists with the unofficially reserved word
“action” in Rails, which is an attribute overridden in the parameters
hash (`params[:action]`) in every controller.

This update changes the name of the model to `action`, and aliases the
object name in the action form to `action_item` so that it can be
safely required in controller params.